+++ test/completion/common.lua
@@ -1870,27 +1870,6 @@ end",
     },
 }
 
-TEST [[
----<??>
-local function f(a, b, c)
-    return a + 1, b .. '', c[1]
-end
-]]
-{
-    {
-        label  = '@param;@return',
-        kind   = define.CompletionItemKind.Snippet,
-        insertText = "\z
-${1:comment}\
----@param a ${2:number}\
----@param b ${3:string}\
----@param c ${4:table}\
----@return ${5:number}\
----@return ${6:string}\
----@return ${7:any}",
-    },
-}
-
 Cared['insertText'] = nil
 
 TEST [[